Python3 vs Python2
I spent a long time installing the dependencies in python2 with pip and trouble shooting a whole host of errors because of this line "python main.py".  I believe your project is written for python3 so this small change may help new comers to jump in faster with pip3 and avoid erroneous trouble shooting.  Best. - X